refactor: split sidebar.rs into sidebar/ module (SRP)
